pyproject.toml
[project]
name = "college-bucket-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.2",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
]
